For loop based bit interleaving
Add functions for integer bit interleaving and deinterleaving in 2, 3
and 4 dimensions to a single integer and to two 64 bit integers for the
4 dimensional case.

Functions use a simple for-loop method (as described here:
http://www.forceflow.be/2013/10/07/morton-encodingdecoding-through-bit-interleaving-implementations/),
could look for possible speed up using other approaches assuming input
and output data types do not hinder selection of constants (i.e. magic
numbers of LUT size).

Includes unit tests and simple benchmark for the 4 dimensional cases
where the dimensions are 16 bit integers and the output is a 64 bit
integer.