Add coordinate conversion functions
Converts from floating point coordinates to integer coordinates. Bith as
an Eigen array or individual coordinates.

Full integer range is used by calculating the fraction of the floating
point range where the point lies.

It seems necessary to increase the size of the floating point coordinate
range slightly to prevent overflow (due to floating point error) when
coordinates are at the edge of the bounds.