Reduce number of task invokations
Reduces the number of OpenMP tasts that are created when hitting the
base case in recursion.

Moves the base case tests to before the tasks are created.

Also updates the maximum recursion depth parameter actually reflect the
depth of the tree ou receive when setting this parameter. Previously it
would not count the root box so the tree was one level deeper than what
the parameter set.